+++ packages/backend-core/src/objectStore/buckets/plugins.ts
@@ -6,7 +6,7 @@ import { Plugin } from "@budibase/types"
 
 // URLS
 
-export function enrichPluginURLs(plugins: Plugin[]): Plugin[] {
+export function enrichPluginURLs(plugins: Plugin[] | undefined): Plugin[] {
   if (!plugins || !plugins.length) {
     return []
   }
