+++ packages/types/src/api/web/index.ts
@@ -1,3 +1,4 @@
+export * from "./application"
 export * from "./analytics"
 export * from "./auth"
 export * from "./user"
